+++ com/main/shipnow/sqlstore/shipnow.go
@@ -46,7 +46,7 @@ func (s *ShipnowStore) Filters(filters []*meta.Filter) *ShipnowStore {
 	return s
 }
 
-func (s *ShipnowStore) Count() (uint64, error) {
+func (s *ShipnowStore) Count() (int, error) {
 	query := s.query().Where(s.preds)
 	return query.Count((*model.ShipnowFulfillment)(nil))
 }
