+++ com/services/affiliate/sqlstore/seller_commission.go
@@ -33,7 +33,7 @@ type SellerCommissionStore struct {
 	filters meta.Filters
 }
 
-func (s *SellerCommissionStore) Count() (uint64, error) {
+func (s *SellerCommissionStore) Count() (int, error) {
 	query := s.query().Where(s.preds)
 	return query.Count((*model.SellerCommission)(nil))
 }
